BUG FIX for one of the cases with quadratic square root
`ExSquareRoot` was broken if any square roots of only external variables are in the input
in addition to the quadratic square root of the next variable in the denominator.

For example, the following
```
ExSquareRoot[sqrt[x]/sqrt[4-4 a[1]+ x a[1]^2], a[1]]
```
was returning incorrect leading singularity `1/sqrt[x]`,
while it is actually unit.